requirements.txt
pip>=21.2.3
python-telegram-bot>=13.7
PyYAML>=6.0
requests>=2.26.0
